Sistemato errore con le classi Modello

+++ src/main/java/MongoDB_Objects/Connessioni/Modello_MongoDB.java
@@ -42,12 +42,19 @@ public class Modello_MongoDB implements Modello_Dao {
 
     public boolean eliminaModello(Integer id_modello) {
         try {
-            Document doc = (Document) this.collection.find(eq("id_modello", id_modello)).first();
-            this.collection.findOneAndDelete(doc);
-            System.out.println("Il modello e' stato eliminato.");
-            return true;
+            Document doc;
+            HashMap<Integer,Modello> modelli = new HashMap<>();
+            modelli = modelliCollection();
+            System.out.println(modelli.getClass());
+            if(modelli.containsKey((int)(long)id_modello)) {
+                doc = (Document) this.collection.find(eq("id_modello", id_modello)).first();
+                this.collection.findOneAndDelete(doc);
+                return true;
+            } else
+                return false;
+
         } catch (Exception e) {
-            System.out.println("Il modello non esiste.");
+            System.out.println("Il modello non esiste ha problemi."+e);
             return false;
         }
     }
@@ -68,7 +75,7 @@ public class Modello_MongoDB implements Modello_Dao {
     public Modello getModello(Integer id_modello) {
         try {
             Document documento = (Document) collection.find(eq("id_modello", id_modello)).first();
-            Modello modello = new Modello(documento.get("id_modello"),(String)documento.get("nome"), (ArrayList) documento.get("attributi"));
+            Modello modello = new Modello(documento.get("id_modello"),(String)documento.get("nome_modello"), (ArrayList) documento.get("attributi"));
             return modello;
         } catch (Exception e) {
             return null;
@@ -79,7 +86,7 @@ public class Modello_MongoDB implements Modello_Dao {
         HashMap<Integer,Modello> modelli = new HashMap();
         List<Document> documents = (List<Document>) collection.find().into(new ArrayList<Document>());
         for(Document documento : documents){
-            Modello modello = new Modello(documento.get("id_modello"),(String)documento.get("nome"), (ArrayList) documento.get("attributi"));
+            Modello modello = new Modello(documento.get("id_modello"),(String)documento.get("nome_modello"), (ArrayList) documento.get("attributi"));
             modelli.put(modello.getId_modello(),modello);
         }
         return modelli;
